version 1.2

+++ migrations/20200123153013_create_shipmentbysap.up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE shipmentbysap (
-    material smallint not null primary key,
+    material integer not null primary key,
     qty integer not null,
     shipment_date date not null default current_date,
     id bigint not null,
